Create utils for tests and enhance readabilty
