--- src/test/axios-for-test.ts
@@ -1,7 +1,0 @@
-import axios from "axios";
-
-const testAxios = axios.create({
-  validateStatus: () => true, // Todos os status codes serão considerados válidos
-});
-
-export default testAxios;
