+++ src/test/test-login.ts
@@ -1,10 +1,7 @@
-import { hash } from "bcrypt-ts";
 import { expect } from "chai";
 import jwt from "jsonwebtoken";
 import "mocha";
-import { start, stop } from "../setup";
-import { prisma } from "../setup-db";
-import axios from "./axios-for-test";
+import axios, { createUser } from "./test-utils";
 
 const test_data = {
   name: "mariana",
@@ -13,33 +10,9 @@ const test_data = {
   birthDate: "2004-10-10",
 };
 
-async function createUser() {
-  return await prisma.user.create({
-    data: {
-      name: test_data.name,
-      email: test_data.email,
-      password: await hash(test_data.password, 8),
-      birthDate: new Date(test_data.birthDate).toISOString(),
-    },
-  });
-}
-
-before(async () => {
-  await start();
-});
-
-after(async () => {
-  await prisma.user.deleteMany();
-  await stop();
-});
-
-afterEach(async () => {
-  await prisma.user.deleteMany();
-});
-
 describe("POST /auth", function () {
   it(`should return a token valid for ${process.env.TOKEN_TIMEOUT} seconds when login is successful`, async function () {
-    const user = await createUser();
+    const user = await createUser(test_data);
     const body = {
       email: test_data.email,
       password: test_data.password,
@@ -65,7 +38,7 @@ describe("POST /auth", function () {
   });
 
   it("should return a token valid for 1 week when login is successful and rememberMe is set to true", async function () {
-    const user = await createUser();
+    const user = await createUser(test_data);
     const body = {
       email: test_data.email,
       password: test_data.password,
@@ -105,7 +78,7 @@ describe("POST /auth", function () {
   });
 
   it("should return an error if the password is incorrect", async function () {
-    await createUser();
+    await createUser(test_data);
     const body = {
       email: test_data.email,
       password: "wrongpassword",
